# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "basePath": "/Login/:Username/:Password",
    "paths": {
        "/Login/:Username/:Password": {
            "get": {
                "description": "do ping",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "ping example",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/Start/{Usecase}/{Appiden}": {
            "get": {
                "description": "do ping",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "ping example",
                "parameters": [
                    {
                        "type": "string",
                        "description": ":Usecase",
                        "name": "Usecase",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": ":Appiden",
                        "name": "Appiden",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/Stop/:Usecase/:Appiden": {
            "get": {
                "description": "do ping",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "ping example",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}